Add reading time and last updated to blog posts
